# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RoomInput {
  /** @minLength 1 */
  hostName: string;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  PRE_RUN: 'PRE_RUN',
  POST_RUN: 'POST_RUN',
} as const;

export type ParticipantInputGender = typeof ParticipantInputGender[keyof typeof ParticipantInputGender];


export const ParticipantInputGender = {
  Male: 'Male',
  Female: 'Female',
} as const;

export type ParticipantInputPreference = typeof ParticipantInputPreference[keyof typeof ParticipantInputPreference];


export const ParticipantInputPreference = {
  P1P2: 'P1P2',
  P3P4: 'P3P4',
  P5P6: 'P5P6',
  NONE: 'NONE',
} as const;

export type ParticipantInputExpertiseItem = typeof ParticipantInputExpertiseItem[keyof typeof ParticipantInputExpertiseItem];


export const ParticipantInputExpertiseItem = {
  NUMBER_0: 0,
  NUMBER_1: 1,
} as const;

export interface ParticipantInput {
  /** @minLength 1 */
  name: string;
  gender: ParticipantInputGender;
  preference: ParticipantInputPreference;
  rank2Preference: ParticipantInputPreference;
  rank3Preference: ParticipantInputPreference;
  /**
     * @minItems 20
     * @maxItems 20
     */
  expertise: ParticipantInputExpertiseItem[];
}

export type ParticipantStatus = typeof ParticipantStatus[keyof typeof ParticipantStatus];


export const ParticipantStatus = {
  UNASSIGNED: 'UNASSIGNED',
  ASSIGNED: 'ASSIGNED',
  NEW_UNASSIGNED: 'NEW_UNASSIGNED',
} as const;

/**
 * @nullable
 */
export type ParticipantAssignedGroup = typeof ParticipantAssignedGroup[keyof typeof ParticipantAssignedGroup] | null;


export const ParticipantAssignedGroup = {
  P1: 'P1',
  P2: 'P2',
  P3: 'P3',
  P4: 'P4',
  P5: 'P5',
  P6: 'P6',
} as const;

export type Participant = ParticipantInput & {
  id: string;
  status: ParticipantStatus;
  /** @nullable */
  assignedGroup: ParticipantAssignedGroup;
};

export type GroupCode = typeof GroupCode[keyof typeof GroupCode];


export const GroupCode = {
  P1: 'P1',
  P2: 'P2',
  P3: 'P3',
  P4: 'P4',
  P5: 'P5',
  P6: 'P6',
} as const;

export interface Group {
  code: GroupCode;
  participantIds: string[];
  maleCount: number;
  femaleCount: number;
}

export interface Room {
  roomCode: string;
  hostName: string;
  hostPassword: string;
  status: RoomStatus;
  participants: Participant[];
  groups: Group[];
  allocationWarnings: string[];
  mcSummary?: MCAllocationSummary | null;
}

export interface MCAllocationSummary {
  iterations: number;
  seed: number;
  best: {
    kpi: KpiResult;
    rankCounts: [number, number, number];
    forcedCount: number;
  };
  metrics: {
    groupParity: Record<GroupCode, number>;
    genderDistribution: { males: number; females: number };
    totalLeaders: number;
  };
  distribution: {
    rank1HitRate: McStats;
    rank2HitRate: McStats;
    genderParityScore: McStats;
    coverageScore: McStats;
    contributionScore: McStats;
    weightedTotal: McStats;
  };
}

export interface McStats {
  mean: number;
  p10: number;
  p50: number;
  p90: number;
}

export interface KpiResult {
  [key: string]: number;
}