# lib/api-zod/src/generated/types/mcAllocationSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KpiResult } from './kpiResult';
import type { McStats } from './mcStats';
import type { GroupCode } from './groupCode';

export interface MCAllocationSummary {
  iterations: number;
  seed: number;
  best: {
    kpi: KpiResult;
    rankCounts: [number, number, number];
    forcedCount: number;
  };
  metrics: {
    groupParity: Record<GroupCode, number>;
    genderDistribution: { males: number; females: number };
    totalLeaders: number;
  };
  distribution: {
    rank1HitRate: McStats;
    rank2HitRate: McStats;
    genderParityScore: McStats;
    coverageScore: McStats;
    contributionScore: McStats;
    weightedTotal: McStats;
  };
}